API is now fixed, API is recieving information.
Due to some poorly named functions, Ben's modal fixes
(which were reasonable) ended up busting the client
talking to the server. I've edited a couple lines and everything
seems to be functioning now. With the new modal design there
are a couple of functions that can be removed. I'll clean
some of those up in the next commit to avoid that kinda thing in
the future.